+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/BaseSpawnerMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
-import insane96mcp.iguanatweaksreborn.module.misc.feature.TempSpawner;
+import insane96mcp.iguanatweaksreborn.module.misc.feature.Spawners;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.BaseSpawner;
@@ -13,14 +13,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(BaseSpawner.class)
 public class BaseSpawnerMixin {
 
-	@Inject(at = @At("HEAD"), method = "serverTick")
+	@Inject(at = @At("HEAD"), method = "serverTick", cancellable = true)
 	private void serverTick(ServerLevel p_151312_, BlockPos p_151313_, CallbackInfo callback) {
-		TempSpawner.onServerTick((BaseSpawner) (Object) this);
+		if (Spawners.onSpawnerServerTick((BaseSpawner) (Object) this))
+			callback.cancel();
 	}
 
-	@Inject(at = @At("HEAD"), method = "clientTick")
-	private void serverTick(Level p_151320_, BlockPos p_151321_, CallbackInfo callback) {
-		TempSpawner.onClientTick((BaseSpawner) (Object) this);
+	@Inject(at = @At("HEAD"), method = "clientTick", cancellable = true)
+	private void clientTick(Level p_151320_, BlockPos p_151321_, CallbackInfo callback) {
+		if (Spawners.onSpawnerClientTick((BaseSpawner) (Object) this))
+			callback.cancel();
 	}
 
 }